+++ src/components/common/organisms/index.ts
@@ -1,3 +1,3 @@
 export { default as Drawer } from './drawer/Drawer';
 export { default as BaseHeader } from './header/BaseHeader';
-export { default as FooterNavi } from './footer/FooterNavi';
+export { default as FooterNavigation } from './footer/FooterNavigation';
